requirements.txt
asgiref==3.9.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.8.0
idna==3.11
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.2
razorpay==2.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
